# Slight Changes for Speed-up