Hi,

I've been working on a project that depends on the inference of Shape-guided Diffusion. I made some slight changes that could reduce the inference time by over 20% when tested on K80 and V100 instances on AWS (the table and screenshots are attached below). The testing script measures the average time of five runs and is also attached to the pull request (`inference_speed_test.py`). I've also checked that the Jupyter notebook generates the same images after my modification to the code.

Could you please verify that the modified code is faithful to your original algorithm, and see if it can help others who would also like to build on your project?

Thank you!

|Instance|Time Before Optimization|Time After Optimization|
|---------------|----------------------------|--------------------------|
|p2.xlarge (Tesla K80)|227s|186s (-22%)|
|p3.2xlarge (Tesla V100)|30.1s|23.2s (-23%)|

Screenshots:
![v100_screenshot](https://github.com/shape-guided-diffusion/shape-guided-diffusion/assets/45331382/e2b322a2-f5e7-4ff1-8582-d4a086d6692c)

![k80_screenshot](https://github.com/shape-guided-diffusion/shape-guided-diffusion/assets/45331382/1773c680-70aa-42ee-9833-3619219b12d4)